Shift around the test names
